separate a statistics table from main menu
