+++ FrontEnd/src/pages/AdminPage/UserProfilesTable/ProfilesStatistics.module.css
@@ -1,0 +1,9 @@
+.description-item-label {
+    max-width: 150px;
+}
+.description-item-content {
+    max-width: 200px;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    white-space: nowrap;
+}
